requirements.txt
python-socketio[client]
discord_webhook

# Only for sending messages
discord.py